Time: 16 ms (88.71%), Space: 56.2 MB (57.26%) - LeetHub

+++ 2665-minimum-time-to-repair-cars/2665-minimum-time-to-repair-cars.java
@@ -5,7 +5,7 @@ class Solution {
         long l = 1;
         long r = 0;
         for(int i = 0; i < ranks.length; i++){
-            r = Math.max(r,(long)ranks[i] * ((long)cars * (long)cars));
+            r = (long)Math.max(r,(long)ranks[i] * ((long)cars * (long)cars));
         }
         while(l <= r){
             long mid = l + (r-l)/2;
